refactor: flatten features folder
