+++ src/app/CanvasContainer.tsx
@@ -1,23 +1,23 @@
 import React, { useEffect, useState } from 'react'
-import { useNavigationStore } from '@/features/navigation/state'
+import { useNavigationStore } from '@/frames/navigationState'
 import type { View } from '@/shared/types'
-import ChatsView from '@/features/chat/ChatsView'
-import FilesView from '@/features/files/FilesView'
-import ReposView from '@/features/repos/ReposView'
-import BranchesView from '@/features/branches/BranchesView'
-import HelpView from '@/features/help/HelpView'
-import CustomersView from '@/features/customers/CustomersView'
-import WeatherView from '@/features/weather/WeatherView'
-import HomeView from '@/features/home/HomeView'
-import InnovationsView from '@/features/innovations/InnovationsView'
-import SettingsView from '@/features/settings/SettingsView'
-import AccountView from '@/features/account/AccountView'
-import NappsView from '@/features/napps/NappsView'
-import TranscludesView from '@/features/transcludes/TranscludesView'
-import MessagesView from '@/features/messages/MessagesView'
-import ProcessesView from '@/features/processes/ProcessesView'
-import AgentsView from '@/features/agents/AgentsView'
-import EventsView from '@/features/events/EventsView'
+import ChatsView from '@/frames/ChatsView'
+import FilesView from '@/frames/FilesView'
+import ReposView from '@/frames/ReposView'
+import BranchesView from '@/frames/BranchesView'
+import HelpView from '@/frames/HelpView'
+import CustomersView from '@/frames/CustomersView'
+import WeatherView from '@/frames/WeatherView'
+import HomeView from '@/frames/HomeView'
+import InnovationsView from '@/frames/InnovationsView'
+import SettingsView from '@/frames/SettingsView'
+import AccountView from '@/frames/AccountView'
+import NappsView from '@/frames/NappsView'
+import TranscludesView from '@/frames/TranscludesView'
+import MessagesView from '@/frames/MessagesView'
+import ProcessesView from '@/frames/ProcessesView'
+import AgentsView from '@/frames/AgentsView'
+import EventsView from '@/frames/EventsView'
 
 const CanvasContainer: React.FC = () => {
   const currentView = useNavigationStore((state) => state.currentView)
